refactor(renderer): extract launcher app helpers
